first attempt at lines plot
