+++ dihedral-plot/11400-1/Abl_dihedral_newsims.npy
